Google-Cloud-Functions/GCS_to_GCS/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage>=2.9.0